+++ src/Lachain.Console/Application.cs
@@ -15,7 +15,7 @@ using Lachain.Core.Network;
 using Lachain.Core.RPC;
 using Lachain.Core.ValidatorStatus;
 using Lachain.Core.Vault;
-using Lachain.Core.Network.FastSync;
+using Lachain.Core.Network.FastSynchronizerBatch;
 using Lachain.Crypto;
 using Lachain.Logger;
 using Lachain.Networking;
