+++ src/Lachain.Core/Network/FastSynchronizerBatch/Downloader.cs
@@ -16,7 +16,7 @@ using Lachain.Storage.Repositories;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public class Downloader : IDownloader
     {
