+++ src/Lachain.Core/Network/FastSynchronizerBatch/FastSyncRepository.cs
@@ -19,7 +19,7 @@ using Lachain.Utility.Serialization;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public class FastSyncRepository : IFastSyncRepository
     {
