+++ src/Lachain.Core/Network/FastSynchronizerBatch/FastSynchronizerBatch.cs
@@ -20,7 +20,7 @@ using Lachain.Utility.Utils;
 using Lachain.Utility.Serialization;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     
     public class FastSynchronizerBatch : IFastSynchronizerBatch
