+++ src/Lachain.Core/Network/FastSynchronizerBatch/HybridQueue.cs
@@ -14,7 +14,7 @@ using Lachain.Utility.Serialization;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public class HybridQueue : IHybridQueue
     {
