+++ src/Lachain.Core/Network/FastSynchronizerBatch/IBlockRequestManager.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public interface IBlockRequestManager
     {
