+++ src/Lachain.Core/Network/FastSynchronizerBatch/IFastSyncRepository.cs
@@ -5,7 +5,7 @@ using Lachain.Storage.Repositories;
 using Lachain.Storage.Trie;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public interface IFastSyncRepository
     {
