+++ src/Lachain.Core/Network/FastSynchronizerBatch/Peer.cs
@@ -2,7 +2,7 @@ using System;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public class Peer : IEquatable<Peer>
     {
