+++ src/Lachain.Core/Network/FastSynchronizerBatch/PeerManager.cs
@@ -9,7 +9,7 @@ using System.Runtime.CompilerServices;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSync
+namespace Lachain.Core.Network.FastSynchronizerBatch
 {
     public class PeerManager
     {
